requirementes.txt
Pillow==9.1.0
pandas==1.4.2
